python/pypi/python-bitcoinlib/requirements.txt
python-bitcoinlib==0.12.2
python-dotenv==1.0.1
